requirements.txt
langchain
streamlit
youtube_transcript_api
pathlib
google-generativeai
python-dotenv

